+++ lib/invoicexpress/models/client.rb
@@ -9,6 +9,7 @@ module Invoicexpress
       element :code, Integer
       element :email, String
       element :address, String
+      element :city, String
       element :postal_code, String
       element :fiscal_id, String
       element :website, String
@@ -19,7 +20,6 @@ module Invoicexpress
       element :balance, Float
       element :country, String
       element :language, String
-      element :city, String
     end
 
   end