+++ lib/invoicexpress/models/invoice.rb
@@ -1,42 +1,11 @@
 require 'invoicexpress/models/client'
+require 'invoicexpress/models/tax'
+require 'invoicexpress/models/item'
+require 'invoicexpress/models/multicurrency'
 
 module Invoicexpress
   module Models
 
-    class Tax < BaseModel
-      include HappyMapper
-
-      tag 'tax'
-      element :id, Integer
-      element :name, String
-      element :value, Float
-      element :region, String
-      element :default_tax, Integer
-    end
-
-    class Item < BaseModel
-      include HappyMapper
-
-      tag 'item'
-      element :id, Integer
-      element :name, String
-      element :description, String
-      element :unit_price, Float
-      element :quantity, Float
-      element :unit, String
-      has_one :tax, Tax
-      element :discount, Float
-    end
-
-    class Items < BaseModel
-      include HappyMapper
-
-      tag 'items'
-      attribute :type, String, :on_save => Proc.new { |value|
-        "array"
-      }
-      has_many :items, Item
-    end
 
     # Fields common to all invoice models, necessary for create/update
     module BaseInvoice
@@ -51,15 +20,26 @@ module Invoicexpress
           element :retention, Float
           element :tax_exemption, String
           element :sequence_id, Integer
-          element :mb_reference, String
+          element :manual_sequence_number, String
 
+          element :mb_reference, String
+          # currency codes
           element :rate, String
           element :currency_code, String
-
           has_one :client, Client
           has_many :items, Item, :on_save => Proc.new { |value|
             Items.new(:items => value)
           }
+        end
+      end
+    end
+
+    # Extra fields for creating credit notes and debit notes
+    module BaseCreditNote
+      def self.included(base)
+        base.class_eval do
+          #The (owner) invoice associated to this credit-note.
+          element :owner_invoice_id, String
         end
       end
     end
@@ -69,7 +49,9 @@ module Invoicexpress
       def self.included(base)
         base.class_eval do
           element :status, String
+          element :archived, String
           element :sequence_number, String
+          element :inverted_sequence_number, String
           element :currency, String
           element :sum, Float
           element :discount, Float
@@ -77,28 +59,37 @@ module Invoicexpress
           element :taxes, Float
           element :total, Float
           element :permalink, String
+          element :saft_hash, String
+          has_one :multicurrency, Multicurrency
         end
       end
 
       def to_core()
-        fields={:date => self.date,
+        fields={
+          :date => self.date,
           :due_date => self.due_date,
           :reference=> self.reference,
           :observations=> self.observations,
           :retention=> self.retention,
           :tax_exemption => self.tax_exemption,
           :sequence_id=> self.sequence_id,
+          :manual_sequence_number=> self.manual_sequence_number,
           :client => self.client,
           :items => self.items,
-          :mb_reference=> self.mb_reference}
+          :currency_code=> self.currency_code,
+          :rate=> self.rate,
+          :mb_reference=> self.mb_reference
+        }
         case self.class.to_s
         when "Invoicexpress::Models::SimplifiedInvoice"
           invoice = Invoicexpress::Models::CoreSimplifiedInvoice.new(fields)
         when "Invoicexpress::Models::CashInvoice"
           invoice = Invoicexpress::Models::CoreCashInvoice.new(fields)
         when "Invoicexpress::Models::CreditNote"
+          fields.merge! owner_invoice_id: self.owner_invoice_id 
           invoice = Invoicexpress::Models::CoreCreditNote.new(fields)
         when "Invoicexpress::Models::DebitNote"
+          fields.merge! owner_invoice_id: self.owner_invoice_id
           invoice = Invoicexpress::Models::CoreDebitNote.new(fields)
         else
           invoice = Invoicexpress::Models::CoreInvoice.new(fields)
@@ -125,12 +116,27 @@ module Invoicexpress
 
     class CoreCreditNote < BaseModel
       include BaseInvoice
+      include BaseCreditNote
       tag 'credit_note'
     end
 
     class CoreDebitNote < BaseModel
       include BaseInvoice
+      include BaseCreditNote
       tag 'debit_note'
+    end
+
+    #added for app.invoicexpress api version
+    class CoreInvoiceReceipt < BaseModel
+      include BaseInvoice
+      tag 'invoice_receipt'
+    end
+
+    #added for app.invoicexpress api version
+    class InvoiceReceipt < BaseModel
+      include BaseInvoice
+      include ExtraInvoice
+      tag 'invoice_receipt'
     end
 
     class SimplifiedInvoice < BaseModel
@@ -153,6 +159,7 @@ module Invoicexpress
 
     class CreditNote < BaseModel
       include BaseInvoice
+      include BaseCreditNote
       include ExtraInvoice
       tag 'credit_note'
 
@@ -161,6 +168,7 @@ module Invoicexpress
 
     class DebitNote < BaseModel
       include BaseInvoice
+      include BaseCreditNote
       include ExtraInvoice
       tag 'debit_note'
 
@@ -185,6 +193,15 @@ module Invoicexpress
       element :total_pages, Integer
       element :total_entries, Integer
       element :per_page, Integer
+    end
+
+    #added for app.invoicexpress api version
+    class InvoiceReceipts < BaseModel
+      include HappyMapper
+
+      tag 'invoice_receipts'
+      has_one :results, InvoiceResult
+      has_many :invoice_receipts, CashInvoice
     end
 
     class CashInvoices < BaseModel
@@ -219,25 +236,7 @@ module Invoicexpress
       element :message, String
     end
 
-    class MessageClient < BaseModel
-      include HappyMapper
 
-      tag 'client'
-      element :email, String
-      element :save, Integer
-    end
-
-    class Message < BaseModel
-      include HappyMapper
-
-      tag 'message'
-      has_one :client, MessageClient
-      element :cc, String
-      element :bcc, String
-      element :subject, String
-      element :body, String
-    end
 
   end
 end
-