+++ lib/invoicexpress/models/multicurrency.rb
@@ -1,0 +1,12 @@
+module Invoicexpress
+  module Models
+    class Multicurrency < BaseModel
+      include HappyMapper
+
+      tag 'multicurrency'
+      element :rate, Float
+      element :currency, String
+      element :total, Float
+    end
+  end
+end