+++ lib/invoicexpress/models/sequence.rb
@@ -6,15 +6,17 @@ module Invoicexpress
       tag 'sequence'
       element :id, String
       element :serie, String
-      element :current_invoice_sequence_id, Integer
-      element :current_invoice_number, Integer
-      element :current_credit_note_sequence_id, Integer
-      element :current_credit_note_number, Integer
-      element :current_debit_note_number, Integer
       element :default_sequence, Boolean
+      element :current_invoice_number, Integer
+      element :current_invoice_sequence_id, Integer
+      element :current_invoice_receipt_number, Integer
+      element :current_invoice_receipt_sequence_id, Integer
       element :current_simplified_invoice_sequence_id, Integer
       element :current_simplified_invoice_number, Integer
+      element :current_credit_note_sequence_id, Integer
+      element :current_credit_note_number, Integer
       element :current_debit_note_sequence_id, Integer
+      element :current_debit_note_number, Integer
       element :current_receipt_sequence_id, Integer
       element :current_receipt_number, Integer
       element :current_shipping_sequence_id, Integer
@@ -27,6 +29,7 @@ module Invoicexpress
       element :current_quote_number, Integer
       element :current_fees_note_sequence_id, Integer
       element :current_fees_note_number, Integer
+      # confirm these 2
       element :current_purchase_order_sequence_id, Integer
       element :current_purchase_order_number, Integer
     end